+++ FinalProject/FinalProject/Bullet.h
@@ -37,7 +37,7 @@ public:
 	Bullet();
 	~Bullet();
 	void Init(D3DXVECTOR3 a_pos, D3DXVECTOR3 a_velocity, PrimObj  a_structpoi, float a_rot, float a_angle, float a_lifespan, int a_damage); // Add parameters
-	void Update(float a_dt); // Parameters?
+	void Update(double a_dt); // Parameters?
 	D3DXVECTOR3 getPos();
 	D3DXVECTOR3 getProspectivePos();
 	int getDamage();
@@ -68,7 +68,7 @@ public:
 	~BulletVec();
 	void Init(ResourceManager& resMan);
 	void Render(DXFrame& DXVid);
-	void Update(float a_dt);
+	void Update(double a_dt);
 	Bullet& GetBullet(int a_index);
 	bool GetActive(int a_index);
 	bool ActivateABullet(D3DXVECTOR3 a_pos, D3DXVECTOR3 a_velocity, int a_type, float a_rot, float a_angle, float a_lifespan, int a_damage);
